1.Interface shouldn't contain concrete implementation, please refer defination of IConsultService. 2.Interface should go under folder interface 3.Best to put this JsonValidatorService under folder concrete, if like rename folder name concrete to services
