+++ Chatbot.Domain/Interface/IJsonValidatorService.cs
@@ -1,0 +1,7 @@
+﻿namespace Chatbot.Domain.Concrete
+{
+    public interface IJsonValidatorService
+    {
+        bool IsValidJson(string jsonString);
+    }
+}
