+++ ChatbotRestAPI.Tests/IntentControllerTests.cs
@@ -1,6 +1,6 @@
 ﻿using Chatbot.Domain.Models;
 using Chatbot.Domain.Ports;
-using Chatbot.Domain.Services;
+using Chatbot.Domain.Concrete;
 using ChatbotRestAPI.Controller;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
